make new file: ppu/register.go
